pyproject.toml
[project]
name = "events-face"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.4.1",
    "drf-yasg>=1.21.10",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "ruff>=0.11.4",
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
